backend/requirements.txt
django~=4.2.0
psycopg2-binary
python-dotenv
google-generativeai